finish sorter implementation
Twitort::Sorter provides some functions for mapping a given user's tweets to a
list of words sorted on frequency of use.  A 'word' is defined by its
construction.  We split a tweet on spaces and then remove any non-word
characters at the beginning and end of the word.  We don't do any special
handling of non-ASCII UTF-8 characters, so these are removed if they appear at
the beginning or end of a word.